service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
pymongo==4.6.0
pinecone==3.0.0
sentence-transformers==2.2.2
google-generativeai==0.3.1
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
jwt